Seperated estimate function for modularity and testing
